requirements.txt
Django==4.0
django-model-utils==4.2.0
django-autoslug==1.9.8
psycopg2-binary==2.9.2

pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.5.2
factory-boy==3.2.1
Faker==11.0.0
